Adds classic data flow liveness analysis
We used to only have a dependency analysis that says which SSA defs/phis
are used at which other SSA defs/phis/jmps. This works great for saying
which flags of our leaky instruction will be used and for not doing
abstract interpretation of non-used defs/phis.

This approach does not work well for saying which flags are live into
a leaky instruction, so adds a quick classic data flow liveness analysis
and log this to the csv file as the column: flags_live_in.