Use config sub name resolution for alerts
This catches the last 3 subs that bap renames so they are resolved
to the name that LLVM sees in the passes (the symbol name in obj file).
This was accidentally removed at some point, but added the call back in
with the new name resolution (just really a refactoring from the config
sub name resolution).

Better warns on some alert filtering/transform passes for visibility.

Adds the --uarch-checker-double-check flag.